Reorganize datasets and metrics in pathology
