+++ monai/apps/pathology/metrics/__init__.py
@@ -9,16 +9,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .data import MaskedInferenceWSIDataset, PatchWSIDataset, SmartCachePatchWSIDataset
-from .handlers import ProbMapProducer
-from .metrics import LesionFROC
-from .transforms.stain.array import ExtractHEStains, NormalizeHEStains
-from .transforms.stain.dictionary import (
-    ExtractHEStainsd,
-    ExtractHEStainsD,
-    ExtractHEStainsDict,
-    NormalizeHEStainsd,
-    NormalizeHEStainsD,
-    NormalizeHEStainsDict,
-)
-from .utils import PathologyProbNMS, compute_isolated_tumor_cells, compute_multi_instance_mask
+from .lesion_froc import LesionFROC
